refactor(gui): rename mixins.gui.{gameoverlay->hud}
